LayoutTests/imported/w3c/web-platform-tests/resize-observer/zoom-expected.txt


FAIL ResizeObserver sizes account for zoom assert_equals: content-rect top should be scaled by zoom expected 5 but got 10
